+++ src/tree/commands/AllCommands.ts
@@ -101,8 +101,15 @@ import {ParallelsCatalogForceRefreshCommand} from "./parallelsCatalogProvider/fo
 import {BuyBusinessLicenseCommonCommand} from "./common/buyBusinessLicense";
 import {ParallelsCatalogPullCatalogManifestCommand} from "./parallelsCatalogProvider/pullCatalogProviderManifest";
 import {BuyProLicenseCommonCommand} from "./common/buyProLicense";
+import {ParallelsCatalogCloseOnboardingCommand} from "./parallelsCatalogProvider/closeOnboarding";
+import {ParallelsCatalogShowAiFaqCommand} from "./parallelsCatalogProvider/showAiFaq";
+import {VagrantInstallCommand} from "./common/installVagrant";
 
-export const AllCommonCommand: CommonCommand[] = [BuyProLicenseCommonCommand, BuyBusinessLicenseCommonCommand];
+export const AllCommonCommand: CommonCommand[] = [
+  BuyProLicenseCommonCommand,
+  BuyBusinessLicenseCommonCommand,
+  VagrantInstallCommand
+];
 
 export const AllVirtualMachineCommands: VirtualMachineCommand[] = [
   AddVmCommand,
@@ -213,5 +220,7 @@ export const AllDevopsRemoteProviderManagementCommands: DevOpsRemoteProviderMana
 export const AllParallelsCommands: ParallelsCatalogCommand[] = [
   ParallelsCatalogRefreshCommand,
   ParallelsCatalogForceRefreshCommand,
-  ParallelsCatalogPullCatalogManifestCommand
+  ParallelsCatalogPullCatalogManifestCommand,
+  ParallelsCatalogCloseOnboardingCommand,
+  ParallelsCatalogShowAiFaqCommand
 ];